(BKR-872) allow host's pe_ver to have precedence over the global pe_ver
This commit introduces a change where we use the global pe_ver if it is
not defined at the host level. Prior to this commit, upgrades would end
up using the global pe_ver for Answer file generation instead of the
host pe_ver. This would work if the global pe_ver was not defined, but
would end up generating the wrong Answer type if the global pe_ver was
defined.