(PE-19169) Extract a FeatureFlags class
to make it easier to test. Flags can be set in the :answers Hash or
picked up from environment variables. In the later case, we need to be
able to preserver them in the :answers hash, if not already present, so
that BeakerAnswers ensures that they are set in the generated pe.conf
file.