(PE-21809) Add aix puppet uninstall function
This PR adds in a way to uninstall puppet on AIX agents.
Whenever an agent is frictionlessly installed (the only
way we support with AIX) or upgraded, the uninstaller script
from the master is copied over to the AIX machine.
The script is called by a method that runs the script then
validates that various puppet related services are no longer
running.