(PE-22724) Add PE managed external postgres support
This PR adds support for external postgres that is installed as part
of the install_pe method.
If a non-infrastructure host has a role 'pe_postgres', then we will
install PE on the normal infrastructure nodes, but put the postgres
database on the role with pe_postgres.
As described in our docs here:
https://puppet.com/docs/pe/2017.3/installing/installing_postgresql.html#install-standalone-pe-postgresql

New methods added in this PR to faciliate the install process:
execute_installer_cmd: Just a wrapper for running the installer command
original_pe_ver and upgrading_to_pe_ver: These are pulled from the
pe_acceptance_tests helper library. They are logic that looks at the
options object to determine if we are preforming an upgrade.