(PE-22724) Add logic to determine installer type for postgres upgrades
Previously if there was an upgrade occuring we'd just consider the install
type as just generic.
With the addition of external postgres support to beaker-pe, that needs to
be handled in its own method. So this PR adds in additional logic for checking
for upgrade if there is a node with the pe_postgres role.