(PE-22724) Fix up the pe_postgres error handling
Previously we were just looking at the error message. This error message
only contained the last 10 lines. In PE 2016.4.x that is enough to ensure
what went wrong. But there was more error handling added after, and so the
message we are attempting to read is higher up and not captured by the error
message.
This PR adds in looking at the lastest install log on the master and grepping
it for the expected error message.