(PE-20250) Add in support for the puppet service debug flag
When frictionlessly installing, if we want to have the puppet service
command run with the debug option we need to pass in a flag.
This PR checks for a host with the option set puppet_service_debug_flag.
If it find that it adds to the frictionless options --puppet-service-debug.
Which pe-repo will then use to turn on debugging for the puppet service command.