(PE-20250) Add version checking for the puppet service debug flag
Previously if we wanted to debug a frictionless install we would
check to see if flag was passed to beaker-pe and then set the flag
on the frictionless install line.
In the case of an upgrade, this caused problem since PE 2016.4.10's
frictionless install would error out if it saw the debug flag passed in.
This PR adds in logic to checking the debug flag to also consider what version
of PE.