(maint) Fix up error handling logic
Previously we grepped the install log when the external postgres job
failed to install on the master. We were looking for a specific error
message that is expected.
If that message isn't in the log, we want to fail with an error message
that makes sense. However, if the grep failed Beaker would exit there
and it left a confusing error message.