(PE-22681) Use packaging_platform instead of platform when set
packaging_platform can be used in beaker-puppet to specify the platform
a pacakge was built on, and when specified in a beaker config should be
the primary source of truth for the platform string. This is the case
for platforms such as redhatfips, which are otherwise treated as an el
platform in beaker.