(PE-23950) Add a manage_puppet_service flag
So that we can continue to manage the puppet agent service per PE-11353
after the next branches are merged into irving. Temporarily this
behavior is still gated by the pe-modules-next flag.

Since pe.conf is now part of the default hiera hierarchy for
puppet_enterprise, we don't need to touch the agent node group.