PE-24016 Update escaped errors with external postgres install
After increasing the rbac timeout, installation on the master is failing
at a different point now with pe-console services restart failure error.
This failure is expected. Added new error to the list of expected errors
to skip past this expected failure  point to continue with the installation
of postgres node.