(maint) Add a timeout to the network resource check
We run the network resource check for each node, for some of our more
complicated testing this can be anywhere from 5 nodes to 10 nodes.
If the curl command times out, this can cause the network resource check
to last for 10-15 minutes.
This PR puts a 10 second timeout on the curl command to try to reduce
the time the network resource check takes.