(BKR-1519) fix login_with_puppet_access_on test
Before, this test would pass in a host value. With the update
to scooter latest in this ticket, we're starting to see that
the deep initialization done here fails on not having details
of the host in the mock given. Since getting to those levels
shouldn't be needed in spec tests for this, I have mocked the
lower levels which are not needed to verify the method code