(maint) Remove special `install_dev_repos_on` code
This commit removes code in `install_dev_repos_on` and routes it to the
`install_puppetlabs_dev_repo` method defined in the FOSSUtils module in
the beaker-puppet repo. Since that `install_dev_repos_on` is used
directly in several projects, that code should be changed in all
affected projects before removing it entirely from this code base.