(PE-27037) Grab fips tarball when using redhatfips-7-x86_64
For testing reasons, the puppet agent team needs the platform tag to remain el-7-x86_64 when using redhatfips-7-x86_64 hosts. This sets the host['dist'] string to use the packaging_platform string rather than the platform string when packaging_platform is redhatfips.

If the platform tag ever changes to redhatfips, this commit should be reverted.