(PE-27048) Modify proxy setup
This changes the setup for using a proxy. It applies the iptables rules to all nodes being tested, puts the dropped connections to osmirror and enterprise at the top of the chain, adds punching through the Platform9 IP range (it is unnecessarily wide so we don't have to go back and fix it if things change slightly), and modifies various curl commands pulling down tarballs when provisioning infra nodes to use the set up proxy.