(PE-27048) Further proxy automation work This commit adds iptable changes for: Test-forge, in order to grab modules for LEI validation. GitHub which is required to clone a repo in order to figure out  what modules to install. We have to punch a hole through the iptables firewall and not use the proxy. This is because GitHub cloning that we need is via ssh, which was too difficult to get working over a http-proxy. Bundler requires udp over port 53, opens that up. Allows hosts to access localhost, a lot of our validation requires that.
Also this commit adds in a hacky work around for what is the root cause of PE-27048.
For now we're not setting the global package manager setting to be configured for a proxy, instead we're configuring each non puppet_enterprise.repo to use the proxy setting.
A proper fix will be to enable a setting in puppet_enteprise::repo::configure to allow us to place proxy=_none_ in the puppet_enteprise.repo package configuration.
Once that is done we'd undo my hacky workaround and just place the proxy setting in the global package manager spot.

(PE-27048) Add SLES proxy logic

This PR adds in logic for SLES pickiness around proxy setup.
We're adding the master in the global proxy settings as no proxy, so
in the future when we fix the root cause of PE-27048 we will have to
remove that setting.