(PE-27048) Finish off the proxy on PE compile master work
This PR adds in support for the proxy job. It also supports all PE master
platforms.
Added in this were:
Ruby bundler/gem works, allow Git to work with proxy, allow testforge access

Those are enough to allow a PE Master, PE CM, and an puppet-agent nodes to all
work together behind a proxy and pass pe_acceptance_tests (with a pending PR in
that repo).