(PE-27048) Undo hacky Yum proxy repo configuration
Previously we configured each Yum repo to use a proxy, and then when
we added the puppet_enterprise repo it would not have any proxy configuration.
This allowed us to test PE with a CM.
Now we're configuring YUM CM repo's with a default of not using a proxy.
So we need to remove the old hacky workaround to confirm the fix works.