(maint) Update PE managed postgres steps
For LEI split upgrades we need to update the pe.conf on the master in order
for a new feature flag to be set.
Previously we alwasy skipped updating the pe.conf on upgrades for all nodes.
For sure on the postgres node if we attempt to update the pe.conf things go
sideways. However in this scenario we only need to update the pe.conf for the
master.
So this PR adds in logic to ignore updating pe.conf on upgrades on postgres node.