(CT-123) Package install with version
PE Clinet Tools pipeline does not know how to downgrade packages.  For
example, if the pe-client-tools package that is to be installed is older
than the pe-client-tools that comes with puppet enterprise, it will not
install the required vesion.

This PR makes an additional check to see if the correct pe-client-tools package was installed.