(maint) Silence batch output in pe-client-tools
Commands that are part of pe-client-tools are executed on Windows with a
batch script.

Without this change, the beaker command output will always include the
called command, which causes inconsistencies between Windows and other
platforms.

Add `@echo off` to the batch script to ensure we don't include the
command in the output.