(maint) Properly determine agent version for puppet-agent 7
The `get_puppet_agent_version` method runs `puppet facts` on the master
to determine the `aio_agent_version` fact. In Puppet versions < 7, the
facts are all nested under a `values` hash, but in Puppet 7, they are at
the top level. That means we need to potentially check two different
locations to find that fact. We can't switch our behavior by Puppet
version, since this is the code that will tell us the Puppet version.