(PE-31503) AIX frictionless install without old TLS
Previously when curling on a AIX host we were using an outdated curl
version that didn't support TLSv3. This meant we had to explictily install
PE with letting in old 1.2 ciphers.
We now have an updated curl and no longer require that. AIX hosts can now
be installed like other platforms, with just the curl -k option.