(PE-32680) Add new GPG when installing Ubuntu/SLES/Debian
This PR adds in the new GPG key when installing older agents. It does
this in two places. If the primary server is Ubuntu, and we're installing PE 2019.8.4
or older, we replace the GPG key in the PE tarball. This will allow Ubuntu to install as
a primary server.
For installing agents of ubuntu/debian/sles, we need to update the GPG key that comes with
pe_repo.