(PE-34061) Specify powershell execution policy during install
beaker-pe could not frictionless install puppet on Window 11, because Windows
client/desktop OSes default to `Restricted` powershell execution policy[1],
and that disables execution of all `.ps1` files (and more).

Now we specify the execution policy during install, which matches the same
options that beaker's `powershell` DSL method uses[2]. We unfortunately can't
use beaker directly because `installer_cmd` needs to return the "command to
execute" instead of executing it directly.

This wasn't an issue in Windows 10 and earlier, because the execution policy
was changed globally in the image.

[1] https://docs.microsoft.com/en-us/powershell/module/microsoft.powershell.core/about/about_execution_policies?view=powershell-7.2#restricted
[2] https://github.com/voxpupuli/beaker/blob/4.36.0/lib/beaker/dsl/wrappers.rb#L41