(PE-34549) Add --waitforlock 1 to agent run post-install
Some VMs run slower than others (looking at you, GCE-backed Windows 11 and Windows Server 2022) and thus take an extra second to shut down the puppet run when stopping the puppet service. This adds the --waitforlock 1 option to the puppet agent -t run after the service shuts down to allow for this extra time and then proceeding with the puppet run, to avoid the command failing.